// src/app/modules/products/shared/enums/action-types.ts
export enum ActionTypes {
  GetProducts = '[Products ] Get All Products ',
  GetProductsSuccess = '[Products ] Get All  Success',
  GetProductsError = '[Products ] Get All  Error',

  GetProduct = '[Products ] Get Product',
  GetProductSuccess = '[Products ] Get Product Success',
  GetProductError = '[Products ] Get Product Error',

  PostProduct = '[Products ] Add Product  ',
  PostProductSuccess = '[Products ]  Add Product Success',
  PostProductError = '[Products ]  Add Product Error',

  UpdateProduct = '[Products ] Update Product  ',
  UpdateProductSuccess = '[Products ]  Update Product Success',
  UpdateProductError = '[Products ]  Update Product Error',

  DeleteProduct = '[Products ] Delete Product  ',
  DeleteProductSuccess = '[Products ]  Delete Product  Success',
  DeleteProductError = '[Products ]  Delete Product  Error',
}